+++ .werft/wipe-devstaging.ts
@@ -1,10 +1,10 @@
 import { werft, exec } from './util/shell';
 import { wipePreviewEnvironment, listAllPreviewNamespaces } from './util/kubectl';
-import  * as fs from 'fs';
-import { sleep } from './util/util';
+import * as fs from 'fs';
+import { deleteExternalIp } from './util/gcloud';
 
 
-async function wipeDevstaging(pathToKubeConfig: string) {
+async function wipePreviewCluster(pathToKubeConfig: string) {
     const namespace_raw = process.env.NAMESPACE;
     const namespaces: string[] = [];
     if (namespace_raw === "<no value>" || !namespace_raw) {
@@ -21,27 +21,6 @@ async function wipeDevstaging(pathToKubeConfig: string) {
     }
 }
 
-async function deleteExternalIp(k3sWsProxyIP: string, namespace: string) {
-    werft.log("wipe", `address describe returned: ${k3sWsProxyIP}`)
-    werft.log("wipe", `found external static IP with matching name ${namespace}, will delete it`)
-
-    const cmd = `gcloud compute addresses delete ${namespace} --region europe-west1 --quiet`
-    let attempt = 0;
-    for (attempt = 0; attempt < 10; attempt++) {
-        let result = exec(cmd);
-        if (result.code === 0 && result.stdout.indexOf("Error") == -1) {
-            werft.log("wipe", `external ip with name ${namespace} and ip ${k3sWsProxyIP} deleted`);
-            break;
-        } else {
-            werft.log("wipe", `external ip with name ${namespace} and ip ${k3sWsProxyIP} could not be deleted, will reattempt`)
-        }
-        await sleep(5000)
-    }
-    if (attempt == 10) {
-        werft.log("wipe", `could not delete the external ip with name ${namespace} and ip ${k3sWsProxyIP}`)
-    }
-}
-
 // if we have "/workspace/k3s-external.yaml" present that means a k3s ws cluster
 // exists, therefore, delete corresponding preview deployment from that cluster too
 // NOTE: Even for a non k3s ws deployment we will attempt to clean the preview.
@@ -50,7 +29,7 @@ async function deleteExternalIp(k3sWsProxyIP: string, namespace: string) {
 async function k3sCleanup() {
     if (fs.existsSync("/workspace/k3s-external.yaml")) {
         werft.log("wipe", "found /workspace/k3s-external.yaml, assuming k3s ws cluster deployment exists, will attempt to wipe it")
-        await wipeDevstaging("/workspace/k3s-external.yaml")
+        await wipePreviewCluster("/workspace/k3s-external.yaml")
         const namespace_raw = process.env.NAMESPACE;
 
         // Since werft creates static external IP for ws-proxy of k3s using gcloud
@@ -61,12 +40,17 @@ async function k3sCleanup() {
         } else {
             werft.log("wipe", `no external static IP with matching name ${namespace_raw} found`)
         }
+    } else {
+        werft.log("wipe", `file /workspace/k3s-external.yaml does not exist, no cleanup for k3s cluster`)
     }
 }
 
+async function devCleanup() {
+    await wipePreviewCluster("")
+}
 // sweeper runs in the dev cluster so we need to delete the k3s cluster first and then delete self contained namespace
-k3sCleanup().then(()=>{
-    wipeDevstaging("")
+k3sCleanup().then(() => {
+    devCleanup()
 })
